==> Python/00_GetData.ipynb <==
!pip install openml --quiet

import openml
import pandas as pd
from openml.datasets import edit_dataset, fork_dataset, get_dataset

import pickle

from google.colab import drive
drive.mount('/content/drive')

import sys,os
sys.path.append('/content/drive/MyDrive/Experiment/')

def get_Xy(id):
  # This is done based on the dataset ID.
  dataset = openml.datasets.get_dataset(id)

  # Print a summary
  print(f"This is dataset '{dataset.name}', the target feature is "
      f"'{dataset.default_target_attribute}'")

  X, y, categorical_indicator, attribute_names = dataset.get_data(target=dataset.default_target_attribute, dataset_format="dataframe")
  return X,y,categorical_indicator, attribute_names

['45058','43892','43890','4535'] ##Credit-g, National-longitudinal-survey-binary, Law-school-admission-binary, Census-Income

X,y,_,_=get_Xy('45058')

X.shape
