==> Python/01_DataProfile.ipynb <==
####Load data

X,y=pd.read_csv(dir+'X.csv'),pd.read_csv(dir+'y.csv')

df=pd.concat([X,y],axis=1)

df

print(f'Orginal dataset shape {df.shape}')

n_samples=df.shape[0]
limitPer = n_samples * 0.5
df = df.dropna(thresh=limitPer, axis=1)
print(f'Drop columns with high missing values {df.shape}')

to_drop=[]
for col in df.columns:
  col_unique_itmes = df[col].nunique()
  ##print single value features
  if (col_unique_itmes == 1):
      to_drop.append(col)
      print(col + " has been added to drop list due to there is single value in this feature")
  elif (col_unique_itmes == df.shape[0]):
      to_drop.append(col)
      print(col + " has been added to drop list as it may be id as values are totally unique")

df = df.drop(to_drop,axis=1)
print(f'Drop columns with single and totally unique values {df.shape}')

####Data profiling - schema detection

def check_if_datetime(x):
    try:
        pd.to_datetime(x)
    except (RuntimeError, TypeError, NameError, IOError, ValueError):
        return False
    else:
        return True


# Define function
def check_if_numeric(x):
    try:
        pd.to_numeric(x)
    except (RuntimeError, TypeError, NameError, IOError, ValueError):
        return False
    else:
        return True


def identify_variable_type(x):
    # Is the column empty? If so, it will be classified as 'NA':
    if (x.dropna().empty == True):
        return "NA"

    if x.dropna().nunique() == len(x)*0.99:
        return "u_manual"

    # Is the variable categorical? We check the number of unique values:
    if ((x.dropna().shape[0] < 3000 and x.dropna().nunique() < 30) or (x.dropna().nunique() < 50 and x.dropna().shape[0] >= 3000)):
        if x.dropna().nunique() == 2:
            return "c_binary"
        else:
            if (check_if_numeric(x) == True):
                return "c_integer"
            else:
                return "c_categorical"
    # If no numbers are present, we classify it as a string:
    elif (x.astype(str).str.contains(r"[0-9]").any() == False):
        return "c_string"
    # We then check if it's numeric, or predominantly numeric with some exceptions:
    elif (check_if_numeric(x) == True):
        if (x.dropna() % 1 == 0).all():
            return "n_integer"
        else:
            return "n_float"
    # next, we check if it's a date or a time, or predominantly datetime with some exceptions:
    elif (check_if_datetime(x) == True):
        return "n_datetime"
    elif (x.astype(str)[x.astype(str).str.contains(r"[0-9]") == False].nunique() < 11 and
          check_if_datetime(x[x.astype(str).str.contains(r"[0-9]") == True]) == True):
        return "n_datetime"
    # If none of the above apply, we classify the variable as string:
    else:
        return "u_manual"


def remove_dtname(text,list_dt):
      for i in list_dt:
          text = text.replace(i, '')
      return text

def prepend(list, str):
    str += '{0}'
    list = [str.format(i) for i in list]
    return (list)


def datatype_detect(data,cat_enc=True):

  df=data.copy()

  column_types=df.apply(identify_variable_type)
  column_types = prepend(column_types, '_')
  column_types=[s[:6] for s in column_types]
  new_column_names = [i + j for i, j in zip(list(df.columns), column_types)]

  df.columns = new_column_names

  col_names=df.columns.to_list()

  c_col=[s for s in col_names if '_c_' in s]
  c_col_index=[df.columns.get_loc(c) for c in c_col]
  n_col=[s for s in col_names if '_n_' in s]
  int_col=[s for s in col_names if 'int' in s]

  lists=['_c_cat','_n_int','_n_flo','_c_bin','_c_str','_c_int']

  c_col=[remove_dtname(i,lists) for i in c_col]
  n_col=[remove_dtname(i,lists) for i in n_col]
  int_col=[remove_dtname(i,lists) for i in int_col]

  cat_dict={}

  if cat_enc:
    data,cat_dict=label_encoding(data,c_col,int_col)

  for col in data.columns:
    if col in int_col:
      data[col] = data[col].astype('int')
      # data[col]=pd.to_numeric(data[col], errors='coerce').fillna(0)
    else:
      data[col] = data[col].astype('float')

  return data,cat_dict,c_col,n_col,c_col_index,int_col

def label_encoding(data,c_col,int_col):

  real=data.copy()
  cat_dict={}
  for i in c_col:
    if i not in int_col:
      real[i]=real[i].astype('category')
      cat_dict[i] = dict(enumerate(real[i].cat.categories))
      real[i]=real[i].cat.codes
      real[i]=real[i].astype('int')

  return real,cat_dict

data,cat_dict,c_col,n_col,c_col_index,int_col=datatype_detect(df)

data.head()

cat_dict

####Statistical Analysis

!pip install pyreadstat --quiet

import pandas as pd
import numpy as np

import pickle
!pip install researchpy --quiet
import researchpy as rp
import scipy.stats as stats

import matplotlib.pyplot as plt
import seaborn as sns

rp.summarize(df[n_col])

rp.summary_cat(df[c_col])

y_name="class"
ss,rs={},{}
for i in n_col:
  summary, results = rp.ttest(group1= df[i][df[y_name] == df[y_name].unique()[1]], group1_name= "Postive",
                              group2= df[i][df[y_name] == df[y_name].unique()[0]], group2_name= "Negative")
  ss[i]=summary
  rs[i]=results


pd.concat(ss)

cb={}
for i in c_col:
  cb[i]=pd.concat([pd.crosstab(df[i],df[y_name]),pd.crosstab(df[i],df[y_name]).apply(lambda r: r/r.sum(), axis=0)],axis=1)

pd.concat(cb)

####Split and save data into 5 seeds

from sklearn.model_selection import train_test_split

for i in range(5):
  X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, test_size=0.2, random_state=i)
  baseline=pd.concat([X_train,y_train],axis=1)
  df_test=pd.concat([X_test,y_test],axis=1)
  baseline.to_csv(dir+'synthetic/seed' + str(i) +'/baseline.csv',index=False)
  df_test.to_csv(dir+'synthetic/seed' + str(i) +'/df_test.csv',index=False)
